feat: golden-fixture regression suite for the scoring engine (#28)
* feat: golden-fixture regression suite for the scoring engine [alertlint-b7f]

End-to-end scoreService composition (classify -> cold-start -> noise ->
threshold -> coverage -> composite -> priority) exercised over four
golden fixtures: kitchen-sink (all states + TH-4 on a self-healing
pile), all-dormant (state, never a score), clean-healthy, and
no-archetype (coverage redistribution + default tier + partial flag).
-update regenerates; a 1-bit golden change fails the suite. 25-run
byte-identical map-order canary.


* fix: prevent cross-architecture FMA fusion in float accumulations

arm64 fuses sum += a*b into FMA, amd64 does not; Go permits the
fusion, so golden fixtures generated on darwin-arm64 failed on the
linux-amd64 runner by one ulp. Split multiply-accumulate into two
statements (assignment forces float64 rounding) in the noise finding
confidence and composite weighted sum; goldens regenerated.


* fix: disable FMA fusion with explicit float64 conversions

Assignment to an intermediate does not reliably disable fusion; the
Go spec's documented mechanism is an explicit conversion. Local arm64
now computes the same unfused values as the amd64 CI runner; goldens
regenerated to the platform-independent results.


---------

+++ internal/score/composite.go
@@ -51,7 +51,9 @@ func Composite(in CompositeInput, cfg Config) CompositeResult {
 	for i, p := range parts {
 		if available[i] {
 			availableWeight += p.weight
-			weightedSum += p.score * p.weight
+			// Explicit conversion disables cross-architecture FMA fusion
+			// (ADR 0005 determinism).
+			weightedSum += float64(p.score * p.weight)
 		} else {
 			res.PartialScore = true
 		}
